fix: filter HUB USB devices
When we connect a device which is indeed a hub that contain additional
devices, we want to add to the menu only the devices connected, this is,
the leaves nodes in the heararchy. For doing that we check the value
for the bDeviceClass property (if it exists), and compare with the value
9 which has been observed a common value for the HUBs.

I would like to know where to find a table that relate that value, to
be confident the above is accurate; the above is based on observations.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>