fix: remove Devices when USB are disconnected
When the USB was disconnected, the DeviceMenuItem was not removed.
It was happening because the getter properties of the extension was
trying to access to information that was not available anymore.

Now the information is stored within the DeviceMenuItem, so when
we receive the disconnect, the new event is accessible, and the
information from the old event is stored in the DeviceMenuItem.

Thanks to this we can properly match the device and remove only
the ones that we are recieving the disconnection event.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>